Vista aceptacion y rechazo de practicas (funcion completa)
> Los datos mostrados ahora son los reales en lugar de solo los id
> Cuando se acepta una practica a un alumno se rechazan automaticamente todas sus demas postulaciones, de modo que jamas pueda estar en dos practicas a la vez

+++ database/migrations/2019_05_28_041429_create_postulaciones_practicas_table.php
@@ -16,7 +16,7 @@ class CreatePostulacionesPracticasTable extends Migration
         Schema::create('postulaciones_practicas', function (Blueprint $table) {
             $table->bigIncrements('id');
             $table->bigInteger('alumnoid')->unsigned();
-            $table->bigInteger('practicaid')->unsigned();
+            $table->Integer('practicaid')->unsigned();
             $table->date('fecha');
             $table->string('estado');
             $table->timestamps();
